+++ src/error.rs
@@ -6,6 +6,9 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
+use std::error;
+use std::fmt;
+
 use simplecss;
 
 use svgtypes;
@@ -17,11 +20,10 @@ use svgtypes::xmlparser::{
 // TODO: split to Dom errors and Parser errors
 
 /// SVG DOM errors.
-#[derive(Fail, Debug)]
+#[derive(Debug)]
 pub enum Error {
     /// If you want to use referenced element inside link attribute,
     /// such element must have a non-empty ID.
-    #[fail(display = "the element must have an id")]
     ElementMustHaveAnId,
 
     /// Linked nodes can't reference each other or itself.
@@ -38,28 +40,19 @@ pub enum Error {
     /// ```text
     /// <linearGradient id="lg1" xlink:href="#lg1"/>
     /// ```
-    #[fail(display = "element crosslink")]
     ElementCrosslink,
 
     /// Parsed document must have an `svg` element.
-    #[fail(display = "the document does not have an SVG element")]
     NoSvgElement,
 
     /// Parsed document must have at least one node.
-    #[fail(display = "the document does not have any nodes")]
     EmptyDocument,
 
     /// *svgdom* didn't support most of the CSS2 spec.
-    #[fail(display = "unsupported CSS at {}", _0)]
     UnsupportedCSS(ErrorPos),
 
     /// Error during parsing of the CSS2.
-    #[fail(display = "invalid CSS at {}", _0)]
     InvalidCSS(ErrorPos),
-
-    /// ENTITY with XML Element data is not supported.
-    #[fail(display = "unsupported ENTITY data at {}", _0)]
-    UnsupportedEntity(ErrorPos),
 
     /// Unexpected close tag.
     ///
@@ -70,40 +63,59 @@ pub enum Error {
     ///     </rect>
     /// </svg>
     /// ```
-    #[fail(display = "opening and ending tag mismatch '{}' and '{}'", _0, _1)]
     UnexpectedCloseTag(String, String),
 
-    /// We don't support a \<paint\> type with a fallback value and a valid FuncIRI.
-    ///
-    /// # Examples
-    ///
-    /// ```text
-    /// <linearGradient id="lg1"/>
-    /// <rect fill="url(#lg1) red"/>
-    /// ```
-    #[fail(display = "valid FuncIRI(#{}) with fallback value is not supported", _0)]
-    UnsupportedPaintFallback(String),
-
-    // TODO: only `use`?
-    /// We don't support `use` elements with a broken filter attribute.
-    #[fail(display = "the 'use' element with a broken filter attribute('#{}') is not supported", _0)]
-    BrokenFuncIri(String),
-
-    /// Failed to find an attribute, which must be set, during post-processing.
-    #[fail(display = "attribute '{}' is missing in the '{}' element", _0, _1)]
-    MissingAttribute(String, String),
-
     /// Error during attribute value parsing.
-    #[fail(display = "{}", _0)]
     SvgTypesError(svgtypes::Error),
 
     /// An XML stream error.
-    #[fail(display = "{}", _0)]
     XmlError(xmlparser::Error),
 
     /// simplecss errors.
-    #[fail(display = "{:?}", _0)]
     CssError(simplecss::Error),
+}
+
+impl fmt::Display for Error {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        match *self {
+            Error::ElementMustHaveAnId => {
+                write!(f, "the element must have an id")
+            }
+            Error::ElementCrosslink => {
+                write!(f, "element crosslink")
+            }
+            Error::NoSvgElement => {
+                write!(f, "the document does not have an SVG element")
+            }
+            Error::EmptyDocument => {
+                write!(f, "the document does not have any nodes")
+            }
+            Error::UnsupportedCSS(pos) => {
+                write!(f, "unsupported CSS at {}", pos)
+            }
+            Error::InvalidCSS(pos) => {
+                write!(f, "invalid CSS at {}", pos)
+            }
+            Error::UnexpectedCloseTag(ref first, ref second) => {
+                write!(f, "opening and ending tag mismatch '{}' and '{}'", first, second)
+            }
+            Error::SvgTypesError(ref e) => {
+                write!(f, "{}", e)
+            }
+            Error::XmlError(ref e) => {
+                write!(f, "{}", e)
+            }
+            Error::CssError(ref e) => {
+                write!(f, "{:?}", e) // TODO: impl Display for simplecss's Error
+            }
+        }
+    }
+}
+
+impl error::Error for Error {
+    fn description(&self) -> &str {
+        "an SVG parsing error"
+    }
 }
 
 impl From<xmlparser::Error> for Error {